управление модальным окном - в отдельный модуль
